Moves detail declarations from Templates.hpp to detail/Templates.hpp
