+++ Library/CMakeLists.txt
@@ -31,6 +31,7 @@ set(PLAYRHO_DETAIL_HDRS
 	"${PLAYRHO_INCLUDE_PREFIX}detail/AABB.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}detail/IndexingNamedType.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}detail/RayCastInput.hpp"
+	"${PLAYRHO_INCLUDE_PREFIX}detail/Templates.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}detail/TypeInfo.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}detail/underlying_type.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}detail/Wider.hpp"
