+++ Library/include/playrho/Templates.hpp
@@ -21,167 +21,9 @@
 #ifndef PLAYRHO_TEMPLATES_HPP
 #define PLAYRHO_TEMPLATES_HPP
 
-#include <algorithm>
-#include <functional>
-#include <iterator>
-#include <limits>
-#include <type_traits>
-#include <utility> // for std::declval
+#include <playrho/detail/Templates.hpp>
 
 namespace playrho {
-
-// Bring standard customization points into the namespace...
-using std::begin;
-using std::cbegin;
-using std::cend;
-using std::data;
-using std::empty;
-using std::end;
-using std::size;
-using std::swap;
-
-namespace detail {
-
-/// @brief Low-level implementation of the is-iterable default value trait.
-template <class T, class = void>
-struct IsIterable : std::false_type {
-};
-
-/// @brief Low-level implementation of the is-iterable true value trait.
-template <class T>
-struct IsIterable<
-    T, std::void_t<decltype(begin(std::declval<T>())), decltype(end(std::declval<T>())),
-                   decltype(++std::declval<decltype(begin(std::declval<T&>()))&>()),
-                   decltype(*begin(std::declval<T>()))>> : std::true_type {
-};
-
-/// @brief Low-level implementation of the is-reverse-iterable default value trait.
-template <class T, class = void>
-struct IsReverseIterable : std::false_type {
-};
-
-/// @brief Low-level implementation of the is-reverse-iterable true value trait.
-template <class T>
-struct IsReverseIterable<
-    T, std::void_t<decltype(rbegin(std::declval<T>())), decltype(rend(std::declval<T>())),
-                   decltype(++std::declval<decltype(rbegin(std::declval<T&>()))&>()),
-                   decltype(*rbegin(std::declval<T>()))>> : std::true_type {
-};
-
-/// @brief Template for determining if the given type is an equality comparable type.
-/// @note This isn't exactly the same as the "EqualityComparable" named requirement.
-/// @see https://en.cppreference.com/w/cpp/named_req/EqualityComparable
-template <class T1, class T2, class = void>
-struct IsEqualityComparable : std::false_type {
-};
-
-/// @brief Template specialization for equality comparable types.
-template <class T1, class T2>
-struct IsEqualityComparable<T1, T2, std::void_t<decltype(T1{} == T2{})>> : std::true_type {
-};
-
-/// @brief Template for determining if the given type is an inequality comparable type.
-template <class T1, class T2, class = void>
-struct IsInequalityComparable : std::false_type {
-};
-
-/// @brief Template specialization for inequality comparable types.
-template <class T1, class T2>
-struct IsInequalityComparable<T1, T2, std::void_t<decltype(T1{} != T2{})>> : std::true_type {
-};
-
-/// @brief Template for determining if the given types are addable.
-template <class T1, class T2 = T1, class = void>
-struct IsAddable : std::false_type {
-};
-
-/// @brief Template specializing for addable types.
-template <class T1, class T2>
-struct IsAddable<T1, T2, std::void_t<decltype(T1{} + T2{})>> : std::true_type {
-};
-
-/// @brief Template for determining if the given types are multipliable.
-template <class T1, class T2, class = void>
-struct IsMultipliable : std::false_type {
-};
-
-/// @brief Template specializing for multipliable types.
-template <class T1, class T2>
-struct IsMultipliable<T1, T2, std::void_t<decltype(T1{} * T2{})>> : std::true_type {
-};
-
-/// @brief Template for determining if the given types are divisable.
-template <class T1, class T2, class = void>
-struct IsDivisable : std::false_type {
-};
-
-/// @brief Template specializing for divisable types.
-template <class T1, class T2>
-struct IsDivisable<T1, T2, std::void_t<decltype(T1{} / T2{})>> : std::true_type {
-};
-
-/// @brief Template for determining if the given type is an "arithmetic" type.
-/// @note In the context of this library, "arithmetic" types are all types which
-///   have +, -, *, / arithmetic operator support.
-template <class T, class = void>
-struct IsArithmetic : std::false_type {
-};
-
-/// @brief Template specialization for valid/acceptable "arithmetic" types.
-template <class T>
-struct IsArithmetic<T, std::void_t<decltype(T{} + T{}), decltype(T{} - T{}), decltype(T{} * T{}),
-                                   decltype(T{} / T{})>> : std::true_type {
-};
-
-/// @brief Has-functor trait template fallback class.
-/// @note This is based off the answer by "jrok" on the <em>StackOverflow</em> website
-///   to the question of: "Check if a class has a member function of a given signature".
-/// @see https://stackoverflow.com/a/16824239/7410358
-template <typename, typename T>
-struct HasFunctor {
-    static_assert(std::integral_constant<T, false>::value,
-                  "Second template parameter needs to be of function type.");
-};
-
-/// @brief Has-functor trait template class.
-/// @note This is based off the answer by "jrok" on the <em>StackOverflow</em> website
-///   to the question of: "Check if a class has a member function of a given signature".
-/// @see https://stackoverflow.com/a/16824239/7410358
-template <typename Type, typename Return, typename... Args>
-struct HasFunctor<Type, Return(Args...)> {
-private:
-    /// @brief Declaration of check function for supporting types given to template.
-    template <typename T>
-    static constexpr auto check(T*) ->
-        typename std::is_same<decltype(std::declval<T>()(std::declval<Args>()...)), Return>::type;
-
-    /// @brief Declaration of check function for non-supporting types given to template.
-    template <typename>
-    static constexpr std::false_type check(...);
-
-    /// @brief Type alias for given template parameters.
-    using type = decltype(check<Type>(nullptr)); // NOLINT(cppcoreguidelines-pro-type-vararg)
-
-public:
-    /// Whether or not the given type has the specified functor.
-    static constexpr auto value = type::value;
-};
-
-/// @brief Gets the maximum size of the given container.
-template <class T>
-constexpr auto max_size(const T& arg) -> decltype(arg.max_size())
-{
-    return arg.max_size();
-}
-
-/// @brief Checks whether the given container is full.
-template <class T>
-constexpr auto IsFull(const T& arg) -> decltype(size(arg) == max_size(arg))
-{
-    return size(arg) == max_size(arg);
-}
-
-} // namespace detail
 
 /// @brief Gets an invalid value for the type.
 /// @tparam T Type to get an invalid value for.
